Make handle_kwargs the identity function for distributions and infer keyword argument handling by argument shape
